Fixed `AccountContentView`
Also fixed all typos with `...ConventView`...

Fixed #704
