+++ src/SharedMauiXamlStylesLibrary.SampleApp/Views/SfTemplatePage.xaml
@@ -2,15 +2,15 @@
 <ContentPage
     xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
     xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-    x:Class="SharedMauiXamlStylesLibrary.SampleApp.Views.TemplatePage"
+    x:Class="SharedMauiXamlStylesLibrary.SampleApp.Views.SfTemplatePage"
     
     xmlns:viewModels="clr-namespace:SharedMauiXamlStylesLibrary.SampleApp.ViewModels"
-    xmlns:contentViews="clr-namespace:AndreasReitberger.Shared.ContentViews;assembly=SharedMauiXamlStylesLibrary"
+    xmlns:contentViews="clr-namespace:AndreasReitberger.Shared.Syncfusion.ContentViews;assembly=SharedMauiXamlStylesLibrary.Syncfusion"
     
     xmlns:icons="clr-namespace:AndreasReitberger.Shared.FontIcons;assembly=SharedMauiXamlStylesLibrary"
     xmlns:iconsSyncfusion="clr-namespace:AndreasReitberger.Shared.Syncfusion.FontIcons;assembly=SharedMauiXamlStylesLibrary.Syncfusion"
     
-    Title="TemplatePage"
+    Title="SfTemplatePage"
     Style="{StaticResource Style.Core.ContentPage.Default}"
     x:DataType="viewModels:TemplatePageViewModel"   
     >
