+++ src/SharedMauiXamlStylesLibrary.SampleApp/Views/SfTemplatePage.xaml.cs
@@ -1,0 +1,12 @@
+using SharedMauiXamlStylesLibrary.SampleApp.ViewModels;
+
+namespace SharedMauiXamlStylesLibrary.SampleApp.Views;
+
+public partial class SfTemplatePage : ContentPage
+{
+    public SfTemplatePage(TemplatePageViewModel viewModel)
+    {
+        InitializeComponent();
+        BindingContext = viewModel;
+    }
+}
