+++ src/SharedMauiXamlStylesLibrary/ContentViews/AccountContentView.xaml
@@ -2,38 +2,39 @@
 <ContentView
     xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
     xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-    x:Class="AndreasReitberger.Shared.Syncfusion.ContentViews.AccountConventView"
+    x:Class="AndreasReitberger.Shared.ContentViews.AccountContentView"
     
-    xmlns:core="clr-namespace:Syncfusion.Maui.Core;assembly=Syncfusion.Maui.Core"
+    xmlns:conventViews="clr-namespace:AndreasReitberger.Shared.ContentViews"
     xmlns:shared="clr-namespace:AndreasReitberger.Shared;assembly=SharedMauiXamlStylesLibrary"
-    xmlns:contentViews="clr-namespace:AndreasReitberger.Shared.Syncfusion.ContentViews"
     >
     <ContentView.Resources>
         <shared:SharedConverters />
     </ContentView.Resources>
     
     <ContentView.ControlTemplate>
-        <ControlTemplate x:DataType="contentViews:AccountConventView">
+        <ControlTemplate x:DataType="conventViews:AccountContentView">
             <Grid
-                Style="{StaticResource Style.Core.Grid.ShellTitleView}"
                 BindingContext="{Binding Source={RelativeSource TemplatedParent}}"
                 IsVisible="{Binding IsVisible}"
                 IsEnabled="{Binding IsEnabled}"
                 Background="{Binding Background}"
                 HeightRequest="{Binding HeightRequest}"
-                ColumnDefinitions="Auto,*,Auto"
+                ColumnDefinitions="*,Auto,Auto"
                 >
-                <core:SfAvatarView
-                    AvatarName="{Binding UsernameText}"
-                    ImageSource="{Binding UserImage}"
-                    Style="{StaticResource Style.Syncfusion.SfAvatarView.Initials}"
-                    />
+                <Grid.Style>
+                    <Style TargetType="Grid" BasedOn="{StaticResource Style.Core.Grid.ShellTitleView}">
+                        <Setter Property="VisualStateManager.VisualStateGroups">
+                            <VisualStateGroupList>
+                                <VisualStateGroup Name="PointOver" />
+                            </VisualStateGroupList>
+                        </Setter>
+                    </Style>
+                </Grid.Style>
                 <ContentPresenter
                     IsVisible="{Binding IsSyncing, Converter={StaticResource BooleanReverseVisibilityConverter}}"
-                    Grid.Column="1"
                     />
                 <HorizontalStackLayout
-                    Grid.Column="2"
+                    Grid.Column="1"
                     IsVisible="{Binding IsSyncing}"
                     >
                     <ActivityIndicator
@@ -49,6 +50,21 @@
                         HorizontalTextAlignment="Start"
                         />
                 </HorizontalStackLayout>
+                <Border
+                    Margin="0"
+                    IsVisible="{Binding IsSyncing, Converter={StaticResource BooleanReverseVisibilityConverter}}"                   
+                    HeightRequest="{Binding HeightRequest}"
+                    WidthRequest="{Binding HeightRequest}"
+                    Grid.Column="2"
+                    Style="{StaticResource Style.Core.Border.Profile}"
+                    StrokeShape="RoundRectangle 25"
+                    Background="{Binding Background}"
+                    >
+                    <Image             
+                        Source="{Binding UserImage}"
+                        Aspect="AspectFit"
+                        />
+                </Border>
             </Grid>
         </ControlTemplate>
     </ContentView.ControlTemplate>
