+++ src/SharedMauiXamlStylesLibrary/SharedMauiXamlStylesLibrary.csproj
@@ -45,8 +45,8 @@
 	</ItemGroup>
 
 	<ItemGroup>
-	  <Compile Update="ContentViews\AccountConventView.xaml.cs">
-	    <DependentUpon>%(Filename)</DependentUpon>
+	  <Compile Update="ContentViews\AccountContentView.xaml.cs">
+	    <DependentUpon>AccountContentView.xaml</DependentUpon>
 	  </Compile>
 	  <Compile Update="ContentViews\ProFeatureLockedContentView.xaml.cs">
 	    <DependentUpon>ProFeatureLockedContentView.xaml</DependentUpon>
@@ -54,17 +54,17 @@
 	  <Compile Update="ContentViews\ProFeatureLockedUpgradeContentView.xaml.cs">
 	    <DependentUpon>ProFeatureLockedUpgradeContentView.xaml</DependentUpon>
 	  </Compile>
-	  <Compile Update="ContentViews\BiometricAuthConventView.xaml.cs">
-	    <DependentUpon>%(Filename)</DependentUpon>
+	  <Compile Update="ContentViews\BiometricAuthContentView.xaml.cs">
+	    <DependentUpon>BiometricAuthContentView.xaml</DependentUpon>
 	  </Compile>
-	  <Compile Update="ContentViews\SwitchHeaderConventView.xaml.cs">
-	    <DependentUpon>SwitchHeaderConventView.xaml</DependentUpon>
+	  <Compile Update="ContentViews\SwitchHeaderContentView.xaml.cs">
+	    <DependentUpon>SwitchHeaderContentView.xaml</DependentUpon>
 	  </Compile>
-	  <Compile Update="ContentViews\SortFilterHeaderConventView.xaml.cs">
-	    <DependentUpon>%(Filename)</DependentUpon>
+	  <Compile Update="ContentViews\SortFilterHeaderContentView.xaml.cs">
+	    <DependentUpon>SortFilterHeaderContentView.xaml</DependentUpon>
 	  </Compile>
-	  <Compile Update="ContentViews\TitleViewAccountConventView.xaml.cs">
-	    <DependentUpon>%(Filename)</DependentUpon>
+	  <Compile Update="ContentViews\TitleViewAccountContentView.xaml.cs">
+	    <DependentUpon>TitleViewAccountContentView.xaml</DependentUpon>
 	  </Compile>
 	  <Compile Update="Resources\Themes\ItemTemplates\CarouselViewItemTemplates.xaml.cs">
 	    <SubType>Code</SubType>
@@ -112,7 +112,7 @@
 	</ItemGroup>
 	
 	<ItemGroup>
-	  <MauiXaml Update="ContentViews\AccountConventView.xaml">
+	  <MauiXaml Update="ContentViews\AccountContentView.xaml">
 	    <Generator>MSBuild:Compile</Generator>
 	  </MauiXaml>
 	  <MauiXaml Update="ContentViews\ProFeatureLockedContentView.xaml">
@@ -121,16 +121,16 @@
 	  <MauiXaml Update="ContentViews\ProFeatureLockedUpgradeContentView.xaml">
 	    <Generator>MSBuild:Compile</Generator>
 	  </MauiXaml>
-	  <MauiXaml Update="ContentViews\BiometricAuthConventView.xaml">
+	  <MauiXaml Update="ContentViews\BiometricAuthContentView.xaml">
 	    <Generator>MSBuild:Compile</Generator>
 	  </MauiXaml>
-	  <MauiXaml Update="ContentViews\SwitchHeaderConventView.xaml">
+	  <MauiXaml Update="ContentViews\SwitchHeaderContentView.xaml">
 	    <Generator>MSBuild:Compile</Generator>
 	  </MauiXaml>
-	  <MauiXaml Update="ContentViews\SortFilterHeaderConventView.xaml">
+	  <MauiXaml Update="ContentViews\SortFilterHeaderContentView.xaml">
 	    <Generator>MSBuild:Compile</Generator>
 	  </MauiXaml>
-	  <MauiXaml Update="ContentViews\TitleViewAccountConventView.xaml">
+	  <MauiXaml Update="ContentViews\TitleViewAccountContentView.xaml">
 	    <Generator>MSBuild:Compile</Generator>
 	  </MauiXaml>
 	  <MauiXaml Update="Resources\Themes\Controls\CollectionView.xaml">
